adjust language detection to exclude non-language folders
When attempting to compile a list of language types, it relies on the
existence of folders inside a locale folder. However, recent
documentation now produces additional folders which trigger the update
script to believe they are languages. Adjusting the update script to
check for the `LC_MESSAGES` folder to consider the folder as a language
type.

Signed-off-by: James Knight <james.d.knight@live.com>